Integrate Tailwind CSS and Material Design
A Tailwind CSS configuration file was created to integrate Material Design's color palette and font families into the project. The necessary libraries '@astrojs/tailwind' and '@udixio/tailwind-material' were included in 'package-lock.json'. This integration helps provide a more modern and standard aesthetic for the UI while utilising the versatility and ease of use of the Tailwind CSS framework. Documentation guide was updated to reflect the installation.